# bun/packages/lockinspiel-frontend/src/openapi/user.json
{"openapi":"3.1.0","info":{"title":"lockinspiel-backend-common","description":"","license":{"name":"","identifier":""},"version":"0.1.0"},"paths":{"/user/profile":{"get":{"tags":["Profile"],"summary":"Get profile","description":"Gets the user profile for the current session","operationId":"get_profile","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"},"example":{"avatar_location":"/user/profile","bio":"I wonder how Alice and Bob are doing","display_name":"John Doe","user_id":"00000000-0000-0000-0000-000000000000"}}}},"4XX":{"description":"It's your fault","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}},"5XX":{"description":"We're having a skill issue","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}}},"security":[{"bearer_jwt":[]}]},"put":{"tags":["Profile"],"summary":"Update profile","description":"Updates the user profile for the current session","operationId":"update_profile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileChangeset"},"example":{"bio":"I wonder how Alice and Bob are doing","display_name":"John Doe"}}},"required":true},"responses":{"200":{"description":"Ok"},"4XX":{"description":"It's your fault","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}},"5XX":{"description":"We're having a skill issue","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}}},"security":[{"bearer_jwt":[]}]},"post":{"tags":["Profile"],"summary":"Create profile","description":"Creates a new user profile for the current session","operationId":"create_profile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsertableUserProfile"},"example":{"bio":"I wonder how Alice and Bob are doing","display_name":"John Doe"}}},"required":true},"responses":{"200":{"description":"Ok"},"4XX":{"description":"It's your fault","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}},"5XX":{"description":"We're having a skill issue","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}}},"security":[{"bearer_jwt":[]}]}},"/user/profile/avatar":{"put":{"tags":["Profile"],"summary":"Replace profile avatar","description":"Returns the URL that should be used to upload an image of the user's new avatar","operationId":"put_avatar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutAvatarQuery"},"example":{"file_extension":"png"}}},"required":true},"responses":{"200":{"description":"Ok","content":{"text/plain":{"schema":{"type":"string"},"example":"/user/profile/avatar/upload"}}},"4XX":{"description":"It's your fault","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}},"5XX":{"description":"We're having a skill issue","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}}},"security":[{"bearer_jwt":[]}]},"delete":{"tags":["Profile"],"summary":"Delete profile avatar","description":"Delete's the user profile's avatar from wherever it is stored","operationId":"delete_avatar","responses":{"200":{"description":"Ok"},"4XX":{"description":"It's your fault","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}},"5XX":{"description":"We're having a skill issue","content":{"text/html":{"schema":{"type":"string"},"example":"<!DOCTYPE html><html><head><meta charset=\"utf8\"></head><body><pre><code>\n   0: Example error\n\nLocation:\n   crates/lockinspiel-backend-common/src/error.rs:125\n\n  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ SPANTRACE ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━\n\n   0: lockinspiel_backend_common::error::placeholder\n      at crates/lockinspiel-backend-common/src/error.rs:121\n   1: tower_http::trace::make_span::request with method=GET uri=/user/openapi/json version=HTTP/1.1\n      at /usr/local/cargo/registry/src/index.crates.io-1949cf8c6b5b557f/tower-http-0.6.8/src/trace/make_span.rs:108</code></pre></body></html>"}}}},"security":[{"bearer_jwt":[]}]}}},"components":{"schemas":{"InsertableUserProfile":{"type":"object","required":["display_name","bio"],"properties":{"bio":{"type":"string"},"display_name":{"type":"string"}}},"PutAvatarQuery":{"type":"object","required":["file_extension"],"properties":{"file_extension":{"type":"string"}}},"UserProfile":{"type":"object","required":["user_id","display_name","bio"],"properties":{"avatar_location":{"type":["string","null"]},"bio":{"type":"string"},"display_name":{"type":"string"},"user_id":{"type":"string","format":"uuid"}}},"UserProfileChangeset":{"type":"object","required":["display_name","bio"],"properties":{"bio":{"type":"string"},"display_name":{"type":"string"}}}},"securitySchemes":{"bearer_jwt":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}